requirements.txt
djangorestframework
#utilities
Pillow
#templates 
django-bootstrap5
django-htmx

#testing
pytest
pytest-django
pytest-cov
pytest-watch

#env
python-dotenv